+++ post-trade/src/main/java/com/multiassetoms/posttrade/model/AverageCostException.java
@@ -1,0 +1,8 @@
+package com.multiassetoms.posttrade.model;
+
+public class AverageCostException extends RuntimeException {
+
+    public AverageCostException(String message) {
+        super(message);
+    }
+}